Improved CSS
This commit further improves the CSS. The Navbar was unstable across the various pages. I have now styled the navbar to be the same size for each sub-page (on desktop only). I have also added some minor styling to the sub-page content so that the page headers (className=subtopic) are centered and spaced from the top of the window. The content beneath these headers on the education and skills page have been centered as well. All of these changes were for the desktop only. I am now moving on to make sure these changes hold up across mobile devices.